Deleted previous bookitem.md and moved loader contents across
